Reorganize task package unit tests
