+++ src/test/java/duke/task/DeadlineTest.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.task;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
@@ -7,9 +7,8 @@ import org.junit.jupiter.api.Test;
 
 import duke.constant.PriorityLevel;
 import duke.exception.DukeException;
-import duke.task.Deadline;
 
-public class DeadlineTest {
+public class    DeadlineTest {
     @Test
     public void createDeadline_noMarkSpecified_unmarkedDeadlineReturned() {
         try {
